backend/pyproject.toml
[project]
name = "python-cards-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.129.0",
    "uvicorn[standard]==0.41.0",
    "sqlalchemy==2.0.46",
    "alembic==1.18.4",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "PyJWT==2.11.0",
    "python-dotenv>=1.1.1",
]
